+++ gzap/apworld/gzdoom/__init__.py
@@ -42,10 +42,10 @@ LauncherComponents.components.append(
 class GZDoomLocation(Location):
     game: str = "gzDoom"
 
-    def __init__(self, options, player: int, loc: DoomLocation, region: Region) -> None:
-        super().__init__(player=player, name=loc.name(), address=loc.id, parent=region)
-        self.access_rule = loc.access_rule(player)
-        if loc.secret and not options.allow_secret_progress.value:
+    def __init__(self, world, loc: DoomLocation, region: Region) -> None:
+        super().__init__(player=world.player, name=loc.name(), address=loc.id, parent=region)
+        self.access_rule = loc.access_rule(world)
+        if loc.secret and not world.options.allow_secret_progress.value:
             self.progress_type = LocationProgressType.EXCLUDED
         else:
             self.progress_type = LocationProgressType.DEFAULT
@@ -192,10 +192,7 @@ class GZDoomWorld(World):
 
             region = Region(map.map, self.player, self.multiworld)
             self.multiworld.regions.append(region)
-            if self.options.pretuning_mode:
-                rule = lambda state: True
-            else:
-                rule = map.access_rule(self)
+            rule = map.access_rule(self)
             menu_region.connect(
                 connecting_region=region,
                 name=f"{map.map}",
@@ -203,9 +200,8 @@ class GZDoomWorld(World):
             for loc in self.pool.locations_in_map(map.map):
                 assert loc.name() not in placed, f"Location {loc.name()} was already placed but we tried to place it again!"
                 placed.add(loc.name())
-                location = GZDoomLocation(self.options, self.player, loc, region)
+                location = GZDoomLocation(self, loc, region)
                 if self.options.pretuning_mode:
-                    location.access_rule = lambda state: True
                     if loc.orig_item:
                         location.place_locked_item(self.create_item(loc.orig_item.name()))
                     elif loc.item: